add execution statistics to workflow list
