+++ packages/cli/src/services/workflow-statistics.service.ts
@@ -1,7 +1,7 @@
 import type { INode, IRun, IWorkflowBase } from 'n8n-workflow';
 import { Service } from 'typedi';
 
-import { StatisticsNames } from '@/databases/entities/workflow-statistics';
+import { StatisticsNames, WorkflowStatistics } from '@/databases/entities/workflow-statistics';
 import { WorkflowStatisticsRepository } from '@/databases/repositories/workflow-statistics.repository';
 import { EventService } from '@/events/event.service';
 import { Logger } from '@/logging/logger.service';
@@ -9,6 +9,7 @@ import { UserService } from '@/services/user.service';
 import { TypedEmitter } from '@/typed-emitter';
 
 import { OwnershipService } from './ownership.service';
+import { WorkflowStatisticsData } from '@/interfaces';
 
 type WorkflowStatisticsEvents = {
 	nodeFetchedData: { workflowId: string; node: INode };
@@ -128,5 +129,43 @@ export class WorkflowStatisticsService extends TypedEmitter<WorkflowStatisticsEv
 		}
 
 		this.eventService.emit('first-workflow-data-loaded', metrics);
+	}
+
+	async getData<
+		C extends 'count' | 'latestEvent',
+		D = WorkflowStatistics[C] extends number ? 0 : null,
+	>(workflowId: string, columnName: C, defaultValue: WorkflowStatistics[C] | D) {
+		const stats = await this.repository.find({
+			select: [columnName, 'name'],
+			where: { workflowId },
+		});
+
+		const data: WorkflowStatisticsData<WorkflowStatistics[C] | D> = {
+			productionSuccess: defaultValue,
+			productionError: defaultValue,
+			manualSuccess: defaultValue,
+			manualError: defaultValue,
+		};
+
+		stats.forEach(({ name, [columnName]: value }) => {
+			switch (name) {
+				case StatisticsNames.manualError:
+					data.manualError = value;
+					break;
+
+				case StatisticsNames.manualSuccess:
+					data.manualSuccess = value;
+					break;
+
+				case StatisticsNames.productionError:
+					data.productionError = value;
+					break;
+
+				case StatisticsNames.productionSuccess:
+					data.productionSuccess = value;
+			}
+		});
+
+		return data;
 	}
 }
